third_party/WebKit/LayoutTests/imported/web-platform-tests/html/webappapis/animation-frames/idlharness-expected.txt

CONSOLE MESSAGE: line 198: callback not yet supported
This is a testharness.js-based test.
PASS Window interface: operation requestAnimationFrame(FrameRequestCallback) 
PASS Window interface: operation cancelAnimationFrame(long) 
PASS Window must be primary interface of window 
FAIL Stringification of window assert_equals: class string of window expected "[object Window]" but got "[object global]"
PASS Window interface: window must inherit property "requestAnimationFrame" with the proper type (0) 
PASS Window interface: calling requestAnimationFrame(FrameRequestCallback) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "cancelAnimationFrame" with the proper type (1) 
PASS Window interface: calling cancelAnimationFrame(long) on window with too few arguments must throw TypeError 
Harness: the test ran to completion.
